Make NodeProperties work with chef-provisioning
